View on Ethereum button for wrapped tokens (#89)
